requirements.txt
fastapi
uvicorn
psycopg2-binary
streamlit
requests
python-multipart
pydantic
